# specs/001-code-learning-platform/contracts/api-spec.yaml
openapi: 3.0.3
info:
  title: AI Code Learning Platform API
  description: |
    REST API for the AI Code Learning Platform - Phase 1 MVP.

    This platform transforms AI-generated code into beginner-friendly learning materials
    through task-based code uploads, automatic analysis, document generation, and Q&A.

    Powered by Google Gemini 3 Flash Preview for fast, cost-effective AI content generation.

    **Authentication**: JWT-based with HTTPOnly cookies
    **Base URL**: `/api/v1`
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: http://localhost:8000/api/v1
    description: Development server
  - url: https://api.codelearning.com/api/v1
    description: Production server
...
paths:
...
  /auth/login:
    post:
      tags: [Authentication]
      summary: Login user
      description: Authenticate user and return JWT tokens in HTTPOnly cookies
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
      responses:
        '200':
          description: Login successful
          headers:
            Set-Cookie:
              description: HTTPOnly cookies with access_token and refresh_token
              schema:
                type: string
                example: access_token=...; HttpOnly; Secure; SameSite=Strict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /projects:
    get:
      tags: [Projects]
      summary: List user's projects
      description: Get all active (non-trashed) projects for authenticated user
      security:
        - cookieAuth: []
      parameters:
        - name: include_trashed
          in: query
          description: Include trashed projects
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Projects retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  projects:
                    type: array
                    items:
                      $ref: '#/components/schemas/Project'
                  total:
                    type: integer
                    example: 5

    post:
      tags: [Projects]
      summary: Create new project
      description: Create a new learning project container
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title:
                  type: string
                  minLength: 1
                  maxLength: 255
                  example: My First Calculator
                description:
                  type: string
                  nullable: true
                  example: Learning to build a calculator with AI help
      responses:
        '201':
          description: Project created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          description: Invalid input
        '401':
          description: Not authenticated
...
  /tasks/{task_id}/document:
    get:
      tags: [Documents]
      summary: Get learning document
      description: Get 7-chapter learning document for task
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/TaskId'
      responses:
        '200':
          description: Document retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LearningDocument'
        '404':
          description: Document not found (may still be generating)
  /tasks/{task_id}/document/status:
    get:
      tags: [Documents]
      summary: Get document generation status
      description: Check status of async document generation
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/TaskId'
      responses:
        '200':
          description: Status retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [pending, in_progress, completed, failed]
                    example: in_progress
                  estimated_time_remaining:
                    type: integer
                    description: Seconds remaining (estimate)
                    example: 120
                  error:
                    type: string
                    nullable: true

  # ==================== Practice ====================
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: access_token
      description: JWT access token in HTTPOnly cookie
  parameters:
...
    TaskId:
      name: task_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        skill_level:
          type: string
          example: Complete Beginner
        created_at:
          type: string
          format: date-time
    Project:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        last_activity_at:
          type: string
          format: date-time
        deletion_status:
          type: string
          enum: [active, trashed]
        trashed_at:
          type: string
          format: date-time
          nullable: true
        task_count:
          type: integer
          description: Total tasks in project
        completed_task_count:
          type: integer
          description: Completed tasks
        progress_percentage:
          type: integer
          description: Overall project progress (0-100)
...
    LearningDocument:
      type: object
      properties:
        id:
          type: string
          format: uuid
        generation_status:
          type: string
          enum: [pending, in_progress, completed, failed]
        content:
          type: object
          description: 7-chapter structured content
          properties:
            chapter1:
              type: object
              properties:
                title:
                  type: string
                summary:
                  type: string
            chapter2:
              type: object
              properties:
                title:
                  type: string
                concepts:
                  type: array
                  items:
                    type: object
            # ... other chapters
        generated_at:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: Invalid input
        detail:
          type: string
          nullable: true
        code:
          type: string
          example: VALIDATION_ERROR